+++ osafw-app/App_Code/fw/FwReports.cs
@@ -314,18 +314,17 @@ public class FwReports
                 }
 
             case "xlsx":
-                var xls_sheets = new Hashtable();
-                var xls_data = new Hashtable();
-                xls_data["rows"] = list_rows;
-                xls_data["headers"] = ((Hashtable)list_rows[0]).Keys.OfType<string>().ToList();
+                {
+                    var out_filename = Utils.isEmpty(render_options["xls_filename"]) ? report_code : (string)render_options["xls_filename"];
+                    // TODO make headers as array of readable values, not the same as fields names
+                    var headers = ((Hashtable)list_rows[0]).Keys.OfType<string>().ToList();
+                    var fields = ((Hashtable)list_rows[0]).Keys.OfType<string>().ToList();
 
-                xls_sheets["Sheet1"] = xls_data;
+                    var filepath = ConvUtils.parsePageNativeExcel(fw, string.Join(",", headers), string.Join(" ", fields), list_rows, out_filename);
 
-                var sheetsOrder = ps.ContainsKey("xls_sheets_order") ? (ArrayList)ps["xls_sheets_order"] : null;
-                var filepath = exportToExcel(xls_sheets, sheetsOrder);
-
-                fw.fileResponse(filepath, report_code + ".xlsx", "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet", "attachment");
-                break;
+                    fw.fileResponse(filepath, report_code + ".xlsx", "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet", "attachment");
+                    break;
+                }
             case "csv":
                 {
                     if (isFileRender())
@@ -398,185 +397,5 @@ public class FwReports
     protected string andNotDeleted(string alias = "")
     {
         return $" and {alias}status<>{db.qi(FwModel.STATUS_DELETED)}";
-    }
-
-    private Dictionary<string, int> GetMaxCharacterWidth(ArrayList rows, List<string> headers) {
-        var maxColWidth = new Dictionary<string, int>();
-
-        foreach (string header in headers) {
-            var cell = header;
-            var cellValue = header;
-
-            maxColWidth.Add(cell, cell.Length < 10 ? 10 : cell.Length);
-        }
-
-        foreach (Hashtable cells in rows) {
-            foreach (string cell in cells.Keys) {
-                var cellValue = Utils.toStr(cells[cell]);
-                var cellTextLength = cellValue.Length;
-
-                if (!maxColWidth.ContainsKey(cell)) {
-                    maxColWidth.Add(cell, cell.Length == 0 ? 50 : cell.Length);
-                }
-
-                if (cellTextLength > maxColWidth[cell]) {
-                    maxColWidth[cell] = cellTextLength;
-                }
-            }
-        }
-
-        return maxColWidth;
-    }
-
-    private DocumentFormat.OpenXml.Spreadsheet.Columns AutoSizeCells(ArrayList rows, List<string> headers) {
-        var maxColWidth = GetMaxCharacterWidth(rows, headers);
-        var columns = new DocumentFormat.OpenXml.Spreadsheet.Columns();
-        double maxWidth  = 10;
-
-        UInt32Value iter = 1;
-        foreach(string item in headers) {
-            var val = maxColWidth[item];
-            var width = ((val * maxWidth + 5) / maxWidth * 256) / 256;
-            DocumentFormat.OpenXml.Spreadsheet.Column col = new DocumentFormat.OpenXml.Spreadsheet.Column();
-            col.BestFit = true;
-            col.Min = iter;
-            col.Max = iter;
-            col.CustomWidth = true;
-            col.Width = (double)width;
-            columns.Append(col);
-            iter += 1;
-        }
-        return columns;
-    }
-
-    private Stylesheet GetStylesheet() {
-        var _Fonts = new DocumentFormat.OpenXml.Spreadsheet.Fonts();
-        _Fonts.Append(new DocumentFormat.OpenXml.Spreadsheet.Font());
-        _Fonts.Append(new DocumentFormat.OpenXml.Spreadsheet.Font(new DocumentFormat.OpenXml.Spreadsheet.Bold()));
-        _Fonts.Append(new DocumentFormat.OpenXml.Spreadsheet.Font(new DocumentFormat.OpenXml.Spreadsheet.Bold()));
-
-        var _Fills = new DocumentFormat.OpenXml.Spreadsheet.Fills(new DocumentFormat.OpenXml.Spreadsheet.Fill());
-        var _Borders = new DocumentFormat.OpenXml.Spreadsheet.Borders(new DocumentFormat.OpenXml.Spreadsheet.Border());
-        _Borders.Append(new DocumentFormat.OpenXml.Spreadsheet.Border());
-
-        var border = new DocumentFormat.OpenXml.Spreadsheet.Border();
-        var topBorder = new DocumentFormat.OpenXml.Spreadsheet.TopBorder();
-        topBorder.Style = BorderStyleValues.Medium;
-        border.TopBorder = topBorder;
-        _Borders.Append(border);
-
-        var _CellFormats = new DocumentFormat.OpenXml.Spreadsheet.CellFormats();
-        _CellFormats.Append(new DocumentFormat.OpenXml.Spreadsheet.CellFormat());
-
-        var cellFormat = new DocumentFormat.OpenXml.Spreadsheet.CellFormat();
-        cellFormat.FontId = 1;
-        cellFormat.FillId = 0;
-        cellFormat.BorderId = 1;
-        _CellFormats.Append(cellFormat); // header
-
-        cellFormat = new DocumentFormat.OpenXml.Spreadsheet.CellFormat();
-        cellFormat.FontId = 1;
-        cellFormat.FillId = 0;
-        cellFormat.BorderId = 2;
-        _CellFormats.Append(cellFormat); // footer
-
-        return new Stylesheet(_Fonts, _Fills, _Borders, _CellFormats);
-    }
-
-    /// <summary>
-    /// Create xls file form hashtable where key is sheet name and value is rows array of hashtables with table data
-    /// </summary>
-    /// <param name="sheetsData"></param>
-    protected string exportToExcel(Hashtable sheetsData, ArrayList sheetsOrder = null) {
-        SpreadsheetDocument spreadSheet = null;
-        var columns = new List<string>();
-
-        var fileName = System.IO.Path.GetTempPath() + Utils.uuid() + ".xlsx";
-
-        try
-        {
-            // create the workbook
-            spreadSheet = SpreadsheetDocument.Create(fileName, SpreadsheetDocumentType.Workbook);
-            spreadSheet.AddWorkbookPart();
-            spreadSheet.WorkbookPart.Workbook = new Workbook();
-            // create the worksheet to workbook relation
-            Sheets sheets = spreadSheet.WorkbookPart.Workbook.AppendChild<Sheets>(new Sheets());
-
-            if (sheetsOrder == null)
-            {
-                sheetsOrder = new ArrayList();
-                foreach(string sheetName in sheetsData.Keys)
-                {
-                    sheetsOrder.Add(sheetName);
-                }
-            }
-
-            UInt32Value sheetNumber = 0;
-            foreach (string sheetName in sheetsOrder)
-            {
-                ArrayList rows = (ArrayList)((Hashtable)sheetsData[sheetName])["rows"];
-                List<string> headers = (List<string>)((Hashtable)sheetsData[sheetName])["headers"];
-                sheetNumber += 1;
-
-                var _SheetData = new SheetData();
-                var _WorksheetPart = spreadSheet.WorkbookPart.AddNewPart<WorksheetPart>();
-
-                Sheet s = new Sheet();
-                s.Id = spreadSheet.WorkbookPart.GetIdOfPart(_WorksheetPart);
-                s.SheetId = sheetNumber;
-                s.Name = sheetName;
-                sheets.AppendChild(s);
-
-                var headerRow = new Row();
-                // create header row
-                foreach (string ColumnName in headers) {
-                    columns.Add(ColumnName);
-                    var cell = new Cell();
-                    cell.StyleIndex = 1;
-                    cell.DataType = CellValues.String;
-                    cell.CellValue = new CellValue(ColumnName);
-                    headerRow.AppendChild(cell);
-                }
-                _SheetData.AppendChild(headerRow);
-
-                // create data rows
-                foreach(Hashtable row in rows) {
-                    var newRow = new Row();
-                    foreach (string col in headers) {
-                        var cell = new Cell();
-                        cell.StyleIndex = 0;
-                        // Set style index 2 for bold text, can be using to highlight totals, etc
-                        //    cell.StyleIndex = 2;
-                        cell.DataType = CellValues.String;
-                        if (row.ContainsKey(col)) {
-                            cell.CellValue = new CellValue(Utils.toStr(row[col]));
-                        }
-
-                        newRow.AppendChild(cell);
-                    }
-                    _SheetData.AppendChild(newRow);
-                }
-
-                _WorksheetPart.Worksheet = new Worksheet();
-                _WorksheetPart.Worksheet.Append(AutoSizeCells(rows, headers));
-                _WorksheetPart.Worksheet.Append(_SheetData);
-            }
-
-            var _StylePart = spreadSheet.WorkbookPart.AddNewPart<WorkbookStylesPart>();
-            _StylePart.Stylesheet = GetStylesheet();
-            _StylePart.Stylesheet.Save();
-
-            // save workbook
-            spreadSheet.WorkbookPart.Workbook.Save();
-        }
-        catch (Exception ex)
-        {
-            throw new UserException(ex.Message);
-        }
-        finally {
-            spreadSheet.Dispose();
-        }
-
-        return fileName;
     }
 }
